backend/requirements.txt
Flask==0.10.1
Flask-OAuth==0.12
Jinja2==2.7.3
MarkupSafe==0.23
PyJWT==0.2.1
Werkzeug==0.10.1
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.19
celery==3.1.17
certifi==14.05.14
distribute==0.6.24
httplib2==0.9
itsdangerous==0.24
kombu==3.0.24
mwoauth==0.2.3
oauth2==1.5.211
oauthlib==0.7.2
pycrypto==2.6.1
pytz==2014.10
redis==2.10.3
requests==2.5.1
requests-oauthlib==0.4.2
simple-crypt==4.0.0
six==1.9.0
sockjs-tornado==1.0.1
tornado==4.1
tornado-redis==2.4.18
wsgiref==0.1.2
